pyproject.toml
[project]
name = "atlas-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.27.0",
    "ipykernel>=6.30.0",
    "pydantic-ai-slim[google]>=0.4.7",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pydantic-ai>=0.4.7",
    "exa-py>=1.0.7",
    "supabase>=2.17.0",
    "functions-framework>=3.5.0",
]
